requirements.txt
alabaster==0.7.12
arrow==1.2.3
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Babel==2.11.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bibtexparser==1.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.2
djangorestframework==3.14.0
docutils==0.17.1
exceptiongroup==1.0.1
fake-useragent==0.1.13
free-proxy==1.0.6
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
pycparser==2.21
Pygments==2.13.0
pymongo==3.12.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
scholarly==1.7.3
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-rtd-theme==1.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.2.4
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
zipp==3.10.0
